requirements.txt
# Core dependencies
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
python-dateutil==2.9.0
tenacity==8.2.3

# AI/ML dependencies
openai==1.35.3
anthropic==0.28.0
tiktoken==0.7.0

# Publishing dependencies
jinja2==3.1.4
tweepy==4.14.0
PyGithub==2.3.0
GitPython==3.1.43

# Pipeline dependencies
dagger-io==0.9.6
pyyaml==6.0.1
cachetools==5.3.3

# Development dependencies
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# Optional dependencies for additional features
aiofiles==23.2.1  # For async file operations
markdown==3.6  # For markdown processing
python-dotenv==1.0.1  # For environment variable management
schedule==1.2.2  # For alternative scheduling
redis==5.0.4  # For distributed caching (optional)
sentry-sdk==2.5.1  # For error tracking (optional)